let clang 6.0.1 compile

+++ src/base/factorization/bigint/sf_mpqs.cc
@@ -507,7 +507,7 @@ namespace LiDIA {
                   }
                 }
                       
-                if(faktor[0] == (char) NULL) {
+                if(faktor[0] == (char)(size_t) NULL) {
                   continue;
                 }
               }
